+++ src/test/java/com/example/demo/helper/MemberSaveHelper.java
@@ -1,4 +1,4 @@
-package com.example.demo.helper.save;
+package com.example.demo.helper;
 
 import com.example.demo.builder.TestMemberBuilder;
 import com.example.demo.domain.member.Member;
