+++ src/test/java/com/example/demo/helper/ReviewSaveHelper.java
@@ -1,4 +1,4 @@
-package com.example.demo.helper.save;
+package com.example.demo.helper;
 
 import com.example.demo.builder.TestReviewBuilder;
 import com.example.demo.domain.cafe.Cafe;
