+++ src/test/java/com/example/demo/helper/StudyOnceSaveHelper.java
@@ -1,4 +1,4 @@
-package com.example.demo.helper.save;
+package com.example.demo.helper;
 
 import java.time.LocalDateTime;
